backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
requests==2.31.0
python-dotenv==1.0.1
pytest-asyncio==0.22.0
httpx==0.25.1
pytest-timeout==2.1.0
pytest-mock==3.14.0
requests-mock==1.11.0   